+++ controllers/contacts/index.js
@@ -1,0 +1,17 @@
+const { ctrlWrapper } = require("../../helpers");
+
+const getAll = require("./getAll");
+const getById = require("./getById");
+const add = require("./add");
+const updateById = require("./updateById");
+const updateStatusContact = require("./updateStatusContact");
+const deleteById = require("./deleteById");
+
+module.exports = {
+  getAll: ctrlWrapper(getAll),
+  getById: ctrlWrapper(getById),
+  add: ctrlWrapper(add),
+  updateById: ctrlWrapper(updateById),
+  updateStatusContact: ctrlWrapper(updateStatusContact),
+  deleteById: ctrlWrapper(deleteById),
+};
