+++ controllers/contacts/updateById.js
@@ -1,0 +1,15 @@
+const { Contact } = require("../../models/contact");
+
+const { HttpError } = require("../../helpers");
+
+
+const updateById = async (req, res) => {
+  const { id } = req.params;
+  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  if (!result) {
+    throw HttpError(404, "Not found");
+  }
+  res.status(200).json(result);
+};
+
+module.exports = updateById;
