+++ controllers/index.js
@@ -1,0 +1,17 @@
+const {
+    getAll,
+    getById,
+    add,
+    updateById,
+    updateStatusContact,
+    deleteById,
+  } = require("./contacts");
+  
+  module.exports = {
+    getAll,
+    getById,
+    add,
+    updateById,
+    updateStatusContact,
+    deleteById,
+  };
